realtek: avoid flooding unknown multicast groups
The goal of IGMP/MLD snooping is to save multicast transmissions, to
forward a multicast packet with a snoopable multicast destination only
to ports where a host has actively signalled interest in it through MLD
reports.

However in the special case of no interested multicast listener we would
so far flood such a multicast packet instead. Which is contrary to our
original goal. This patch changes this by avoiding forwarding snoopable
multicast packets with no listeners altogether.

We need to make sure that on the other hand we still keep flooding
multicast packets with an unsnoopable multicast destination, as stated
by RFC4541, section 2.1.2.2 (224.0.0.x) + section 3 (ff02::1).
Therefore when disabling the VLAN L2 unknown multicast flooding setting
we are adding virtual entries for these unsnoopable addresses to ensure
we keep flooding these.

To not flood snoopable but unknown multicast addresses is also the default
behaviour of a pure software Linux bridge.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>