realtek: forward multicast packets to multicast router ports
With this patch a multicast packet is not only forwarded to detected
multicast listeners, but also to detected multicast routers or manually
configured multicast router ports.

For one thing, this now allows using layer 3 multicast routers behind a
switch port.

Even more importantly though this forwarding is required by
RFC4541, section 2.1.2.1. Otherwise not only for routable multicast
packets but also for link-local multicast packet loss might occur
if the elected multicast querier is not on this device.
And hence IPv6 unicast would break, too, as it relies on link-local
multicast for its IPv6 Neighbor Discovery.

Fixes: cde31976e375 ("realtek: Add support for Layer 2 Multicast")
Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>